:recycle: (js) restructure js files
