+++ app/javascript/packs/admin.js
@@ -3,6 +3,6 @@ require('@rails/ujs').start();
 require('@rails/activestorage').start();
 require('channels');
 
-const componentRequireContext = require.context('apps/admin', true);
+const componentRequireContext = require.context('src/apps/admin', true);
 const ReactRailsUJS = require('react_ujs');
 ReactRailsUJS.useContext(componentRequireContext);
