+++ app/javascript/packs/application.js
@@ -3,6 +3,6 @@ require('@rails/ujs').start();
 require('@rails/activestorage').start();
 require('channels');
 
-const componentRequireContext = require.context('apps/application', true);
+const componentRequireContext = require.context('src/apps/application', true);
 const ReactRailsUJS = require('react_ujs');
 ReactRailsUJS.useContext(componentRequireContext);
