+++ app/javascript/src/apps/admin/graphql/documents.ts
