+++ app/javascript/src/apps/admin/graphql/index.ts
