+++ app/javascript/src/apps/admin/graphql/mutations/logout.gql
