+++ app/javascript/src/apps/admin/pages/index.ts
