+++ app/javascript/src/apps/application/components/transaction-detail.tsx
