+++ app/javascript/src/apps/application/components/utxo-detail.tsx
