+++ app/javascript/src/apps/application/graphql/documents.ts
