+++ app/javascript/src/apps/application/graphql/index.ts
