+++ app/javascript/src/apps/application/graphql/mutations/create-multisig-request.gql
