+++ app/javascript/src/apps/application/graphql/mutations/verify-multisig-request.gql
