+++ app/javascript/src/apps/application/graphql/queries/assets.gql
