+++ app/javascript/src/apps/application/graphql/queries/current-group.gql
