+++ app/javascript/src/apps/application/graphql/queries/current-user.gql
