+++ app/javascript/src/apps/application/graphql/queries/multisig-account.gql
