+++ app/javascript/src/apps/application/graphql/queries/multisig-accounts.gql
