+++ app/javascript/src/apps/application/graphql/queries/multisig-transactions.gql
