+++ app/javascript/src/apps/application/pages/account-list/index.ts
