+++ app/javascript/src/apps/application/pages/account-new/index.ts
