+++ app/javascript/src/apps/application/pages/account/components/balance-tab.tsx
@@ -1,5 +1,5 @@
-import { UtxoDetail } from '@/application/components';
-import { MultisigUtxo } from '@/graphql/application';
+import { UtxoDetail } from '@application/components';
+import { MultisigUtxo } from '@application/graphql';
 import {
   Avatar,
   Button,
