+++ app/javascript/src/apps/application/pages/account/components/index.ts
