+++ app/javascript/src/apps/application/pages/account/index.ts
