+++ app/javascript/src/apps/application/pages/index.ts
