+++ app/javascript/src/shared/client.ts
