+++ app/javascript/src/shared/index.ts
