+++ app/javascript/src/shared/interfaces.ts
