+++ config/webpack/customized_config.js
@@ -3,7 +3,30 @@ const path = require("path");
 module.exports = {
   resolve: {
     alias: {
-      "@": path.resolve(__dirname, "..", "..", "app/javascript/apps"),
+      '@admin': path.resolve(
+        __dirname,
+        '..',
+        '..',
+        'app/javascript/src/apps/admin',
+      ),
+      '@application': path.resolve(
+        __dirname,
+        '..',
+        '..',
+        'app/javascript/src/apps/application',
+      ),
+      '@images': path.resolve(
+        __dirname,
+        '..',
+        '..',
+        'app/javascript/src/images',
+      ),
+      '@shared': path.resolve(
+        __dirname,
+        '..',
+        '..',
+        'app/javascript/src/shared',
+      ),
     }
   }
 };
